requirements.txt
browser-use
loguru
openai
pydantic
python-dotenv
